+++ src/components/blogs/BlogReadPage.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState, useCallback, useContext } from "react";
 import { useParams, Link, useNavigate } from "react-router-dom";
-import { getDatabase, ref, get } from 'firebase/database';
+import { getDatabase, ref, get,remove,update } from 'firebase/database';
 import moment from "moment";
 import DOMPurify from "dompurify";
 import { marked } from "marked";
@@ -13,6 +13,7 @@ import { signOut } from "firebase/auth";
 import { ThemeContext } from '../../App';
 import { toast } from "react-toastify";
 import { auth } from "../../firebase/auth";
+import { FaTrash } from "react-icons/fa";
 
 const BlogReadPage = () => {
   const { id } = useParams();
@@ -20,8 +21,30 @@ const BlogReadPage = () => {
   const [menuOpen, setMenuOpen] = useState(false);
   const [isLoggedIn, setLogin] = useState(false);
   const { theme, toggleTheme } = useContext(ThemeContext);
-  const navigate = useNavigate();
+  const userId = localStorage.getItem('userUid');
 
+  const navigate = useNavigate();
+  let [ids,setIds]=useState([])
+  const [user, setUser] = useState(null);
+  useEffect(() => {
+    const fetchUserData = async () => {
+      try {
+        const db = getDatabase();
+        const userRef = ref(db, 'users/' + userId);
+        
+        const userSnap = await get(userRef);
+        if (userSnap.exists()) {
+          setUser(userSnap.val());
+        } else {
+          console.log('No user data available');
+        }
+      } catch (error) {
+        console.error('Error fetching user data:', error);
+      }
+    };
+
+    fetchUserData();
+  }, [userId]);
   useEffect(() => {
     if (localStorage.getItem('login')) {
       setLogin(true);
@@ -60,6 +83,7 @@ const BlogReadPage = () => {
         if (snapshot.exists()) {
           const data = snapshot.val();
           setBlog(data);
+          console.log('fddddddddddd',data)
         } else {
           console.log('No data available');
         }
@@ -78,7 +102,60 @@ const BlogReadPage = () => {
   const createMarkup = (content) => {
     return { __html: DOMPurify.sanitize(marked(content)) };
   };
+  const handleDelete = async (id) => {
+    console.log(user)
+      let isUser=true
+      user.articleCreated.split(',').map((data)=>{
+        if(data==id){
+          isUser=true
+        }
+      })
+      if(!isUser){
+        return
+      }
+      const db = getDatabase();
+        // Remove the article from the articles collection
+       let a= await remove(ref(db, 'articles/' + id));
+        // Update the user's articleCreated list
+        const updatedArticles = user.articleCreated
+          .split(',')
+          .filter((id) => id !== id)
+          .join(',');
+    console.log(db)
+        await update(ref(db, 'users/' + userId), {
+          articleCreated: updatedArticles,
+        });
+    
+        toast.success("Blog Deleted Successfully!! 🚀", {
+          className: "toast-message",
+        });
+        const articlesRef = ref(db, 'articles');
+        const snapshot = await get(articlesRef);
+    
+        if (snapshot.exists()) {
+          const data = snapshot.val();
+          console.log(data)
+          if(Object.values(data).length==0){
+            navigate('/blogs')
 
+    return
+          }
+          const blogsArray = Object.values(data).map(blog => ({
+            title: blog.title,
+            date: new Date(blog.createdAt).toLocaleDateString(),
+            summary: stripMarkdown(blog.content.substring(0, 100)) + '...',
+            tags: blog.tags,
+            author: blog.author,
+            link: `/blogs/${blog.id}`
+          }));
+          navigate('/blogs')
+        } else {
+          navigate('/blogs')
+
+          console.log('No data available');
+        }
+      
+    };
   return (
     <>
       <nav className={`navbar fixed`}>
@@ -127,11 +204,13 @@ const BlogReadPage = () => {
                 <p className="blog-date">{moment(blog.createdAt).fromNow()}</p>
               </div>
             </div>
+
             <div className="blog-tags">
               {blog.tags.map((tag, tagIndex) => (
                 <span key={tagIndex} className="blog-tag">{tag}</span>
               ))}
             </div>
+              {blog.userId==user.id && <div className="flex justify-end" style={{display:'flex',justifyContent:"end"}}>{<FaTrash size={'2rem'} onClick={()=>handleDelete(blog.id)}/>}</div>}
           </div>
           <div className="blog-content" dangerouslySetInnerHTML={createMarkup(blog.content)}></div>
         </div>
